+++ tests/support/rabook_pipeline_fixture.h
@@ -165,7 +165,7 @@ static mem_disk_t s_disk = {};
 /* RAM block backend (4 MiB -> FAT16 via ra8_fs_format) */
 /* -------------------------------------------------------------------------- */
 
-static inline ra8_err_t mem_read(void* ctx, uint32_t lba, uint32_t count, uint8_t* buf)
+static inline ra8_err_t mem_read(void* ctx, uint64_t lba, uint32_t count, uint8_t* buf)
 {
   mem_disk_t* disk = (mem_disk_t*)ctx;
   if (lba + count > disk->block_count) {
@@ -177,7 +177,7 @@ static inline ra8_err_t mem_read(void* ctx, uint32_t lba, uint32_t count, uint8_
   return k_ra8_ok;
 }
 
-static inline ra8_err_t mem_write(void* ctx, uint32_t lba, uint32_t count, const uint8_t* buf)
+static inline ra8_err_t mem_write(void* ctx, uint64_t lba, uint32_t count, const uint8_t* buf)
 {
   mem_disk_t* disk = (mem_disk_t*)ctx;
   if (lba + count > disk->block_count) {
@@ -189,7 +189,7 @@ static inline ra8_err_t mem_write(void* ctx, uint32_t lba, uint32_t count, const
   return k_ra8_ok;
 }
 
-static inline ra8_err_t mem_capacity(void* ctx, uint32_t* block_count, uint32_t* block_size)
+static inline ra8_err_t mem_capacity(void* ctx, uint64_t* block_count, uint32_t* block_size)
 {
   mem_disk_t* disk = (mem_disk_t*)ctx;
   *block_count     = disk->block_count;
